Add time control to groups
Close groups after max_time parameter so groups cannot get stuck forever

+++ server/grouphug-server/src/config.rs
@@ -18,7 +18,7 @@ pub struct Electrum {
 
 #[derive(Deserialize)]
 pub struct GroupConf {
-    pub max_time: usize,
+    pub max_time: i64,
     pub max_size: usize,
 }
